+++ common/routes/Lylive/components/RelatedNews.js
@@ -46,8 +46,9 @@ class RelatedNews extends PureComponent {
           { this.props.newsList.map((item, index) => {
             return (
               <Link className={css(styles.newsBlock)}
+                data-on='click' data-event-category='lylive' data-event-action='relation-news'
                 key={item.sn}
-                to={item.parseUrl}>
+                to={`${item.parseUrl}?from=lylist`}>
                 <div className={css(styles.newsBlock)}>
                   <img src={item.MainPhoto && item.MainPhoto.thumbnail} className={css(styles.img)} />
                   <div className={css(styles.bottom)}>