+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 ## [Unreleased]
 
 ### Added
+- 加上閒置頁面, 90 秒以後出現 @wb
 - 增加 autoTrack @esbb48
+
+### Fixed
+- 修改閒置頁面圖片新聞的樣式，防止新聞圖被壓縮 @wb
 
 ## 1.0.37 - 2017-11-09
 - 更換 ads.txt 內容 @wb
@@ -15,7 +19,6 @@
 
 ## 1.0.37 - 2017-11-02
 ### Added
-- 加上閒置頁面, 90 秒以後出現 @wb
 - 依照業務需求加入 ads.txt @wb
 - 增加文中 DFP `/5799246/Nownews_all_article_300x250_artm` 廣告 @esbb48